requirements.txt
# Core Framework
langchain==0.1.20
langchain-community==0.0.38

# Vector Database (local, embedded)
chromadb

# Embeddings (local)
sentence-transformers==2.5.1
torch==2.2.0

# LLM Integration (Groq uses OpenAI-compatible API)
openai==1.12.0

# Document Processing
pypdf==4.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Web Interface
streamlit==1.31.1

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.3

# Development
jupyter==1.0.0
ipykernel==6.29.2